--- src/main/java/mission/controller/HomeController.java
@@ -1,12 +1,0 @@
-package mission.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HomeController {
-    @GetMapping("/ping")
-    public String ping() {
-        return "ok";
-    }
-}
