+++ src/main/java/mission/post/application/PostController.java
@@ -1,10 +1,11 @@
-package mission.post;
+package mission.post.application;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import mission.post.dto.CreatePostRequest;
 import mission.post.dto.PostCreateResponse;
 import mission.post.dto.UpdatePostRequest;
+import mission.post.business.PostService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
