+++ src/main/java/mission/post/repository/PostRepository.java
@@ -1,5 +1,6 @@
-package mission.post;
+package mission.post.repository;
 
+import mission.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
