+++ src/main/java/mission/user/business/UserService.java
@@ -1,0 +1,8 @@
+package mission.user.business;
+
+import mission.user.dto.SignUpRequest;
+import mission.user.dto.SignUpResponse;
+
+public interface UserService {
+    SignUpResponse signup(SignUpRequest signUpRequest);
+}
