+++ src/main/java/mission/user/domain/User.java
@@ -1,4 +1,4 @@
-package mission.user;
+package mission.user.domain;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
