+++ src/main/java/mission/user/implement/UserServiceImpl.java
@@ -1,17 +1,21 @@
-package mission.user;
+package mission.user.implement;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import mission.user.domain.User;
 import mission.user.dto.SignUpRequest;
 import mission.user.dto.SignUpResponse;
+import mission.user.repository.UserRepository;
+import mission.user.business.UserService;
 import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class UserService {
+public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
 
+    @Override
     @Transactional
     public SignUpResponse signup(SignUpRequest signUpRequest) {
         if(userRepository.existsByEmail(signUpRequest.email())){
